+++ static/js/message_edit.js
@@ -10,6 +10,7 @@ import * as blueslip from "./blueslip";
 import * as channel from "./channel";
 import * as compose from "./compose";
 import * as compose_actions from "./compose_actions";
+import * as compose_ui from "./compose_ui";
 import * as composebox_typeahead from "./composebox_typeahead";
 import * as condense from "./condense";
 import * as confirm_dialog from "./confirm_dialog";
@@ -473,12 +474,12 @@ function edit_message(row, raw_content) {
                 currently_editing_messages.get(rows.id(row)).listeners = listeners;
             }
             composebox_typeahead.initialize_compose_typeahead(edit_id);
-            compose.handle_keyup(null, $(edit_id).expectOne());
+            compose_ui.handle_keyup(null, $(edit_id).expectOne());
             $(edit_id).on("keydown", function (event) {
-                compose.handle_keydown(event, $(this).expectOne());
+                compose_ui.handle_keydown(event, $(this).expectOne());
             });
             $(edit_id).on("keyup", function (event) {
-                compose.handle_keyup(event, $(this).expectOne());
+                compose_ui.handle_keyup(event, $(this).expectOne());
             });
             break;
         }
